requirements.txt
# requirements.txt
pypetkitapi
sqlalchemy
psycopg2-binary
python-dotenv